LayoutTests/inspector/unit-tests/target-manager.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script>
let worker;

function createWorker() {
    worker = new Worker("../worker/resources/worker-1.js");
}

function terminateWorker() {
    worker.terminate();
}

function test()
{
    function typeString(type) {
        switch (type) {
        case WebInspector.Target.Type.Main:
            return "Main";
        case WebInspector.Target.Type.Worker:
            return "Worker";
        default:
            return "Unknown";
        }
    }

    function dumpTargets() {
        for (let target of WebInspector.targets)
            InspectorTest.log(`Target - ${typeString(target.type)} - ${target.displayName}`);
    }


    let suite = InspectorTest.createAsyncSuite("TargetManager");

    suite.addTestCase({
        name: "TargetManager.MainTarget",
        description: "We should always have the main target.",
        test(resolve, reject) {
            InspectorTest.assert(WebInspector.targets === WebInspector.targetManager.targets);
            InspectorTest.expectEqual(WebInspector.targets.size, 1, "Targets list should always start out with the main target.");
            InspectorTest.expectEqual([...WebInspector.targets][0], WebInspector.mainTarget, "Target list should always contain the main target.");
            InspectorTest.expectNotNull(WebInspector.mainTarget.executionContext, "Main target should have an ExecutionContext.");
            InspectorTest.expectEqual(WebInspector.mainTarget.RuntimeAgent, RuntimeAgent, "Main target should have the global RuntimeAgent.");
            dumpTargets();
            resolve();
        }
    });

    suite.addTestCase({
        name: "TargetManager.WorkerTarget.Create",
        description: "Creating a Worker should create a new Worker Target.",
        test(resolve, reject) {
            InspectorTest.evaluateInPage("createWorker()");
            WebInspector.targetManager.singleFireEventListener(WebInspector.TargetManager.Event.TargetAdded, (event) => {
                let target = event.data.target;
                InspectorTest.assert(target instanceof WebInspector.Target);
                InspectorTest.expectEqual(target.type, WebInspector.Target.Type.Worker, "Added Target should have Worker type.");
                InspectorTest.expectNotNull(target.executionContext, "Added Target should have an ExecutionContext.");
                InspectorTest.expectNotNull(target.RuntimeAgent, "Added Target should have a RuntimeAgent.");
                InspectorTest.expectNotEqual(target.RuntimeAgent, RuntimeAgent, "Added Target RuntimeAgent should not be the global RuntimeAgent.");
                dumpTargets();
                resolve();
            });
        }
    });

    suite.addTestCase({
        name: "TargetManager.WorkerTarget.Remove",
        description: "Creating a Worker should create a new Worker Target.",
        test(resolve, reject) {
            InspectorTest.evaluateInPage("terminateWorker()");
            WebInspector.targetManager.singleFireEventListener(WebInspector.TargetManager.Event.TargetRemoved, (event) => {
                let target = event.data.target;
                InspectorTest.assert(target instanceof WebInspector.Target);
                InspectorTest.expectEqual(target.type, WebInspector.Target.Type.Worker, "Removed Target should have Worker type.");
                dumpTargets();
                resolve();
            });
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test for TargetManager and other global WebInspector.Target objects.</p>
</body>
</html>
